requirements.txt
opencv-python
numpy
img2pdf
Pillow
scikit-image
flask
python-magic-bin
reportlab
werkzeug